% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartaUpdatingFunctions.R
\name{spartaNimWeights}
\alias{spartaNimWeights}
\title{Fit a full or reduced state-space model using either MCMC or SMC methods}
\usage{
spartaNimWeights(
  model,
  MCMCconfiguration = NULL,
  pfType = "bootstrap",
  pfControl = list(saveAll = TRUE, smoothing = FALSE),
  nParFiltRun = NULL,
  latent,
  block = FALSE,
  mcmc = TRUE
)
}
\arguments{
\item{model}{A NIMBLE model object, typically representing a state space model or a hidden Markov model.}

\item{MCMCconfiguration}{A list specifying the parameterisation for nimble's runMCMC function.}

\item{pfType}{The type of particle filter algorithm to fit. Defualts to 'bootstrap'.}

\item{pfControl}{A list specifying different control options for the particle filter. Options are described in the details section below.}

\item{nParFiltRun}{Number of particles for the particle filtering algorithm.}

\item{latent}{A character specifying which variable in the state-space model is the latent variable.}

\item{block}{A logical value indicating whether to block the model parameters and assign a random-walk block sampler.}

\item{mcmc}{A logical value indicating whether to fit the model with 'mcmc' or 'particle filter'. Defaults to TRUE.}
}
\description{
Fit a given NIMBLE state space model.
}
\details{

}
\examples{
## For illustration only.
stateSpaceCode <- nimbleCode({
  x0 ~ dnorm(0, var = 1)
  x[1] ~ dnorm(a* x0, var = 1)
  y[1] ~ dnorm(x[1]*c, var = .5)
  for(t in 2:10){
    x[t] ~ dnorm(a * x[t-1], var = 1)
    y[t] ~ dnorm(x[t]*c, var = .5)
  }
})

model <- nimbleModel(code = exampleCode, data = list(y = rnorm(10)),
                     inits = list(x0 = 0, x = rnorm(10)))

 spartaNimWeights(model,
 MCMCconfiguration = list(target = c('a', 'c'),
n.iter = 1000,
n.chains = 2,
n.burnin = 100,
n.thin = 2),
latent = "x",
mcmc = TRUE)


}
\seealso{
Other smc update methods: 
\code{\link{buildAuxiliaryFilterUpdate}},
\code{\link{buildBootstrapFilterUpdate}},
\code{\link{spartaNimUpdates}()},
\code{\link{updateUtils}()}
}
\author{
Kwaku Peprah Adjei
}
\concept{smc update methods}
